Add end to end test verifying setting relayer approval with signatures
